Move game headers into Common
