+++ ChairLoader/ChairLoader.vcxproj
@@ -184,7 +184,7 @@
       <PrecompiledHeader>Use</PrecompiledHeader>
       <PrecompiledHeaderFile>pch.h</PrecompiledHeaderFile>
       <LanguageStandard>stdcpp17</LanguageStandard>
-      <AdditionalIncludeDirectories>$(ProjectDir);</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(ProjectDir);../Common;</AdditionalIncludeDirectories>
       <RuntimeLibrary>MultiThreadedDLL</RuntimeLibrary>
     </ClCompile>
     <Link>
@@ -205,7 +205,7 @@
       <PrecompiledHeader>NotUsing</PrecompiledHeader>
       <PrecompiledHeaderFile>pch.h</PrecompiledHeaderFile>
       <LanguageStandard>stdcpp17</LanguageStandard>
-      <AdditionalIncludeDirectories>$(ProjectDir);%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(ProjectDir);../Common;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
     </ClCompile>
     <Link>
       <SubSystem>Windows</SubSystem>
@@ -227,7 +227,7 @@
       <PrecompiledHeader>NotUsing</PrecompiledHeader>
       <PrecompiledHeaderFile>pch.h</PrecompiledHeaderFile>
       <LanguageStandard>stdcpp17</LanguageStandard>
-      <AdditionalIncludeDirectories>$(ProjectDir);$(SolutionDir)boost_1_77_0;$(SolutionDir)detours;$(SolutionDir)imgui-1.87;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(ProjectDir);../Common;$(SolutionDir)boost_1_77_0;$(SolutionDir)detours;$(SolutionDir)imgui-1.87;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
     </ClCompile>
     <Link>
       <SubSystem>Windows</SubSystem>
@@ -239,10 +239,43 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="ArkAbilityLibrary.h" />
-    <ClInclude Include="ArkBasicTypes.h" />
-    <ClInclude Include="ArkEntityArchetypeLibrary.h" />
-    <ClInclude Include="ArkEntityClassLibrary.h" />
+    <ClInclude Include="..\Common\mem.h" />
+    <ClInclude Include="..\Common\preyDllObjects.h" />
+    <ClInclude Include="..\Common\PreyFunctions.h" />
+    <ClInclude Include="..\Common\Prey\ArkAbilityLibrary.h" />
+    <ClInclude Include="..\Common\Prey\ArkBasicTypes.h" />
+    <ClInclude Include="..\Common\Prey\ArkEntityArchetypeLibrary.h" />
+    <ClInclude Include="..\Common\Prey\ArkEntityClassLibrary.h" />
+    <ClInclude Include="..\Common\Prey\ArkEnums.h" />
+    <ClInclude Include="..\Common\Prey\CryCore\BaseTypes.h" />
+    <ClInclude Include="..\Common\Prey\CryCore\Compiler\MSVCspecific.h" />
+    <ClInclude Include="..\Common\Prey\CryCore\platform.h" />
+    <ClInclude Include="..\Common\Prey\CryInput\BaseInput.h" />
+    <ClInclude Include="..\Common\Prey\CryInput\IHardwareMouse.h" />
+    <ClInclude Include="..\Common\Prey\CryInput\IInput.h" />
+    <ClInclude Include="..\Common\Prey\CryInput\Keyboard.h" />
+    <ClInclude Include="..\Common\Prey\CryMath\Cry_Math.h" />
+    <ClInclude Include="..\Common\Prey\CryMath\Cry_Matrix33.h" />
+    <ClInclude Include="..\Common\Prey\CryMath\Cry_Matrix44.h" />
+    <ClInclude Include="..\Common\Prey\CryMath\Cry_MatrixDiag.h" />
+    <ClInclude Include="..\Common\Prey\CryMath\Cry_Vector2.h" />
+    <ClInclude Include="..\Common\Prey\CryMath\Cry_Vector3.h" />
+    <ClInclude Include="..\Common\Prey\CryMath\Cry_Vector4.h" />
+    <ClInclude Include="..\Common\Prey\CryMath\NumberVector.h" />
+    <ClInclude Include="..\Common\Prey\CryNetwork\ISerialize.h" />
+    <ClInclude Include="..\Common\Prey\CryRenderer\CommonResource.h" />
+    <ClInclude Include="..\Common\Prey\CryRenderer\DeviceResources.h" />
+    <ClInclude Include="..\Common\Prey\CryRenderer\IRenderer.h" />
+    <ClInclude Include="..\Common\Prey\CryRenderer\ITexture.h" />
+    <ClInclude Include="..\Common\Prey\CryRenderer\Renderer.h" />
+    <ClInclude Include="..\Common\Prey\CryRenderer\Texture.h" />
+    <ClInclude Include="..\Common\Prey\CrySystem\HardwareMouse.h" />
+    <ClInclude Include="..\Common\Prey\CrySystem\ISystem.h" />
+    <ClInclude Include="..\Common\Prey\CrySystem\ITimer.h" />
+    <ClInclude Include="..\Common\Prey\CrySystem\Profiling.h" />
+    <ClInclude Include="..\Common\Prey\CryThreading\IJobManager.h" />
+    <ClInclude Include="..\Common\proc.h" />
+    <ClInclude Include="..\Common\VTableHook.h" />
     <ClInclude Include="ChairLoader.h" />
     <ClInclude Include="ChairloaderGui.h" />
     <ClInclude Include="ChairloaderGUIConsole.h" />
@@ -250,53 +283,26 @@
     <ClInclude Include="ChairloaderGUILog.h" />
     <ClInclude Include="ChairloaderGUIPlayerManager.h" />
     <ClInclude Include="EntityUtils.h" />
-    <ClInclude Include="ArkEnums.h" />
     <ClInclude Include="ChairLoaderImGui.h" />
     <ClInclude Include="framework.h" />
     <ClInclude Include="GUIUtils.h" />
     <ClInclude Include="ImGui\imgui_widget_flamegraph.h" />
-    <ClInclude Include="mem.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="PerfOverlay.h" />
-    <ClInclude Include="PreyFunctions.h" />
-    <ClInclude Include="preyDllObjects.h" />
-    <ClInclude Include="Prey\CryCore\BaseTypes.h" />
-    <ClInclude Include="Prey\CryCore\Compiler\MSVCspecific.h" />
-    <ClInclude Include="Prey\CryCore\platform.h" />
-    <ClInclude Include="Prey\CryInput\BaseInput.h" />
-    <ClInclude Include="Prey\CryInput\IHardwareMouse.h" />
-    <ClInclude Include="Prey\CryInput\IInput.h" />
-    <ClInclude Include="Prey\CryMath\Cry_Math.h" />
-    <ClInclude Include="Prey\CryMath\Cry_Matrix33.h" />
-    <ClInclude Include="Prey\CryMath\Cry_Matrix44.h" />
-    <ClInclude Include="Prey\CryMath\Cry_MatrixDiag.h" />
-    <ClInclude Include="Prey\CryMath\Cry_Vector2.h" />
-    <ClInclude Include="Prey\CryMath\Cry_Vector3.h" />
-    <ClInclude Include="Prey\CryMath\Cry_Vector4.h" />
-    <ClInclude Include="Prey\CryNetwork\ISerialize.h" />
-    <ClInclude Include="Prey\CryRenderer\CommonResource.h" />
-    <ClInclude Include="Prey\CryRenderer\DeviceResources.h" />
-    <ClInclude Include="Prey\CryRenderer\IRenderer.h" />
-    <ClInclude Include="Prey\CryRenderer\ITexture.h" />
-    <ClInclude Include="Prey\CryRenderer\Renderer.h" />
-    <ClInclude Include="Prey\CryRenderer\Texture.h" />
-    <ClInclude Include="Prey\CrySystem\HardwareMouse.h" />
-    <ClInclude Include="Prey\CrySystem\ISystem.h" />
-    <ClInclude Include="Prey\CrySystem\ITimer.h" />
-    <ClInclude Include="Prey\CrySystem\Profiling.h" />
-    <ClInclude Include="Prey\CryThreading\IJobManager.h" />
-    <ClInclude Include="proc.h" />
     <ClInclude Include="Profiler.h" />
-    <ClInclude Include="VTableHook.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="..\Common\mem.cpp" />
+    <ClCompile Include="..\Common\PreyDllObjects.cpp" />
+    <ClCompile Include="..\Common\PreyFunctions.cpp" />
+    <ClCompile Include="..\Common\proc.cpp" />
+    <ClCompile Include="..\Common\VTableHook.cpp" />
     <ClCompile Include="ChairLoader.cpp" />
     <ClCompile Include="ChairloaderGui.cpp" />
     <ClCompile Include="EntityUtils.cpp" />
     <ClCompile Include="ChairLoaderImGui.cpp" />
     <ClCompile Include="dllmain.cpp" />
     <ClCompile Include="ImGui\imgui_widget_flamegraph.cpp" />
-    <ClCompile Include="mem.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
@@ -306,11 +312,7 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Library|x64'">Create</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="PerfOverlay.cpp" />
-    <ClCompile Include="PreyFunctions.cpp" />
-    <ClCompile Include="PreyDllObjects.cpp" />
-    <ClCompile Include="proc.cpp" />
     <ClCompile Include="Profiler.cpp" />
-    <ClCompile Include="VTableHook.cpp" />
   </ItemGroup>
   <ItemGroup>
     <Text Include="codeArchive.txt" />
