+++ Common/Prey/CryCore/BaseTypes.h
