+++ Common/Prey/CryCore/platform.h
