+++ Common/Prey/CryInput/IHardwareMouse.h
