+++ Common/Prey/CryMath/Cry_Math.h
