+++ Common/Prey/CryRenderer/CommonResource.h
