+++ Common/Prey/CryRenderer/DeviceResources.h
