+++ Common/Prey/CryRenderer/IRenderer.h
