+++ Common/Prey/CryRenderer/ITexture.h
