+++ Common/Prey/CryRenderer/Renderer.h
