+++ Common/Prey/CryRenderer/Texture.h
