+++ Common/VTableHook.cpp
