+++ Common/VTableHook.h
