+++ Common/mem.cpp
